Create bitstamp-daily.py
This is used to download the daily data from the HTTP source. I set it up to run as a cron job, so for your machine you may have to do some tweaking if you use it. Its not entirely done, but it works, i was just going to add a couple design things to it that are not related to performance. Ill come back and update it when im done.